+++ core/arch/arm/plat-sam/sam_sfr.c
@@ -7,6 +7,7 @@
 
 #include <io.h>
 #include <kernel/dt.h>
+#include <kernel/dt_driver.h>
 #include <matrix.h>
 #include <mm/core_memprot.h>
 #include <mm/core_mmu.h>
