+++ core/drivers/atmel_wdt.c
@@ -10,6 +10,7 @@
 #include <io.h>
 #include <kernel/delay.h>
 #include <kernel/dt.h>
+#include <kernel/dt_driver.h>
 #include <kernel/pm.h>
 #include <matrix.h>
 #include <sama5d2.h>
