+++ core/drivers/imx_lpuart.c
@@ -8,6 +8,7 @@
 #include <io.h>
 #include <keep.h>
 #include <kernel/dt.h>
+#include <kernel/dt_driver.h>
 #include <util.h>
 
 #define STAT		0x14
