+++ core/drivers/pl011.c
@@ -7,6 +7,7 @@
 #include <io.h>
 #include <keep.h>
 #include <kernel/dt.h>
+#include <kernel/dt_driver.h>
 #include <stdlib.h>
 #include <trace.h>
 #include <types_ext.h>
