+++ core/drivers/rstctrl/stm32_rstctrl.c
@@ -10,6 +10,7 @@
 #include <io.h>
 #include <kernel/delay.h>
 #include <kernel/dt.h>
+#include <kernel/dt_driver.h>
 #include <kernel/panic.h>
 #include <mm/core_memprot.h>
 #include <stm32_util.h>
