+++ core/include/kernel/dt.h
@@ -62,14 +62,6 @@ struct dt_device_match {
 	const void *compat_data;
 };
 
-enum dt_driver_type {
-	DT_DRIVER_NOTYPE,
-	DT_DRIVER_UART,
-	DT_DRIVER_CLK,
-	DT_DRIVER_RSTCTRL,
-	DT_DRIVER_I2C,
-};
-
 /*
  * DT_MAP_AUTO: Uses status properties from device tree to determine mapping.
  * DT_MAP_SECURE: Force mapping for device to be secure.
@@ -81,42 +73,7 @@ enum dt_map_dev_directive {
 	DT_MAP_NON_SECURE
 };
 
-/*
- * dt_driver_probe_func - Callback probe function for a driver.
- *
- * @fdt: FDT base address
- * @nodeoffset: Offset of the node in the FDT
- * @compat_data: Data registered for the compatible that probed the device
- *
- * Return TEE_SUCCESS on successful probe,
- *	TEE_ERROR_DEFER_DRIVER_INIT if probe must be deferred
- *	TEE_ERROR_ITEM_NOT_FOUND when no driver matched node's compatible string
- *	Any other TEE_ERROR_* compliant code.
- */
-typedef TEE_Result (*dt_driver_probe_func)(const void *fdt, int nodeoffset,
-					   const void *compat_data);
-
-#if defined(CFG_DT)
-/*
- * Driver instance registered to be probed on compatible node found in the DT.
- *
- * @name: Driver name
- * @type: Drive type
- * @match_table: Compatible matching identifiers, null terminated
- * @driver: Driver private reference or NULL
- * @probe: Probe callback (see dt_driver_probe_func) or NULL
- */
-struct dt_driver {
-	const char *name;
-	enum dt_driver_type type;
-	const struct dt_device_match *match_table; /* null-terminated */
-	const void *driver;
-	TEE_Result (*probe)(const void *fdt, int node, const void *compat_data);
-};
-
-#define DEFINE_DT_DRIVER(name) \
-		SCATTERED_ARRAY_DEFINE_PG_ITEM(dt_drivers, struct dt_driver)
-
+#ifdef CFG_DT
 /*
  * Find a driver that is suitable for the given DT node, that is, with
  * a matching "compatible" property.
@@ -347,10 +304,4 @@ static inline int fdt_get_reg_props_by_name(const void *fdt __unused,
 }
 
 #endif /* !CFG_DT */
-
-#define for_each_dt_driver(drv) \
-	for (drv = SCATTERED_ARRAY_BEGIN(dt_drivers, struct dt_driver); \
-	     drv < SCATTERED_ARRAY_END(dt_drivers, struct dt_driver); \
-	     drv++)
-
 #endif /* KERNEL_DT_H */
